refactor. pages 폴더 내 구조화, 스타일드 컴포넌트 코드 분리
