--- src/pages/Detail.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-import DetailView from '../components/Detail/DetailView';
-
-export default function Detail() {
-  return (
-    <>
-      <DetailView />
-    </>
-  );
-}
