+++ src/pages/Detail/Detail.tsx
@@ -1,0 +1,9 @@
+import DetailView from '../../components/Detail/DetailView';
+
+export default function Detail() {
+  return (
+    <>
+      <DetailView />
+    </>
+  );
+}
