--- src/pages/Main.tsx
@@ -1,90 +1,0 @@
-import styled from 'styled-components';
-import { dbService } from '../common/firebase';
-import { useEffect, useState } from 'react';
-import { getDocs, collection } from 'firebase/firestore';
-import ArtistView from '../components/Main/ArtistView';
-import { useAppDispatch } from '../hooks/useRedux';
-import { saveArtists } from '../redux/modules/ArtistsSlice';
-import { ArtistsTypes } from '../redux/modules/ArtistsSlice';
-
-export default function Main() {
-  const [artists, setArtists] = useState<any[]>([]);
-  const dispatch = useAppDispatch();
-
-  const getArtist = async () => {
-    const querySnapshot = await getDocs(collection(dbService, 'artists'));
-    const artist: ArtistsTypes[] = [];
-
-    querySnapshot.forEach((doc) => {
-      const newArtist: ArtistsTypes = {
-        id: doc.id,
-        ...doc.data(),
-      };
-
-      artist.push(newArtist);
-    });
-
-    setArtists(artist);
-    dispatch(saveArtists(artist));
-  };
-
-  useEffect(() => {
-    getArtist();
-  }, []);
-
-  return (
-    <MainPageWrapper>
-      <MainPageIntroWrapper>
-        <MainPageIntroArea />
-      </MainPageIntroWrapper>
-
-      <MainPageComponentsWrapper>
-        <MainPageArtistBoxArea>
-          {artists.map((item: any) => (
-            <ArtistView item={item} key={item.id} />
-          ))}
-        </MainPageArtistBoxArea>
-      </MainPageComponentsWrapper>
-    </MainPageWrapper>
-  );
-}
-
-// MainPage
-const MainPageWrapper = styled.div`
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: column;
-  min-width: 200px;
-  min-height: 100vh;
-`;
-
-// MainPageIntro
-const MainPageIntroWrapper = styled.div`
-  background-color: #000;
-
-  height: 300px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-const MainPageIntroArea = styled.div`
-  background: url('/assets/bounce.png') no-repeat center;
-
-  width: 500px;
-  height: 150px;
-`;
-
-// MainPageComponents
-const MainPageComponentsWrapper = styled.div`
-  background-color: #f7f7f7;
-
-  padding: 0 160px;
-  position: relative;
-`;
-const MainPageArtistBoxArea = styled.div`
-  flex-wrap: wrap;
-  display: flex;
-  justify-content: center;
-  margin: 0 auto;
-  padding: 70px 50px;
-`;
