+++ src/pages/Mypage/style.ts
@@ -1,0 +1,23 @@
+import styled from 'styled-components';
+
+export const MyPageWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  background-color: #f7f7f7;
+`;
+
+export const Container = styled.div`
+  margin-left: 50px;
+  margin-right: 50px;
+`;
+
+export const Section = styled.section`
+  min-height: 300px;
+  margin: 3rem 0 4rem;
+`;
+
+export const SectionTitle = styled.h1`
+  font-size: 24px;
+  margin-left: 1rem;
+`;
