+++ src/shared/Router.tsx
@@ -1,14 +1,14 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import Footer from '../components/Layout/Footer';
 import Header from '../components/Layout/Header';
-import ScrollToTopButton from '../components/ScrollToTopButton';
-import Detail from '../pages/Detail';
-import Error from '../pages/Error';
-import Login from '../pages/Login';
-import Main from '../pages/Main';
-import Media from '../pages/Media';
-import Mypage from '../pages/Mypage';
-import SignUp from '../pages/SignUp';
+import ScrollToTopButton from '../components/ScrollTopTopButton/ScrollToTopButton';
+import Detail from '../pages/Detail/Detail';
+import Error from '../pages/Error/Error';
+import Login from '../pages/Login/Login';
+import Main from '../pages/Main/Main';
+import Media from '../pages/Media/Media';
+import Mypage from '../pages/Mypage/Mypage';
+import SignUp from '../pages/SignUp/SignUp';
 import ScrollUpto from '../components/ScrollUpTo';
 
 const Router = () => {
